Rename opentracing-globaltracer to opentracing-util module.
Introduced the `io.opentracing.util` package.
